[refactor] remove duplicated UserDto UserMapper
