--- src/main/java/pl/wsb/fitnesstracker/training/internal/UserMapper.java
@@ -1,16 +1,0 @@
-package pl.wsb.fitnesstracker.training.internal;
-
-import org.springframework.stereotype.Component;
-import pl.wsb.fitnesstracker.user.api.User;
-
-@Component("trainingUserMapper")
-class UserMapper { // todo: user mapper from user/internal package
-
-    UserDto toDto(User user) {
-        return new UserDto(user.getId(),
-                user.getFirstName(),
-                user.getLastName(),
-                user.getBirthdate(),
-                user.getEmail());
-    }
-}
