+++ src/main/java/pl/wsb/fitnesstracker/user/api/UserDto.java
@@ -1,4 +1,4 @@
-package pl.wsb.fitnesstracker.user.internal;
+package pl.wsb.fitnesstracker.user.api;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.annotation.Nullable;
@@ -6,7 +6,7 @@ import jakarta.annotation.Nullable;
 import java.time.LocalDate;
 
 //todo; move it to api also UserMapper
-record UserDto(@Nullable Long id, String firstName, String lastName,
+public record UserDto(@Nullable Long id, String firstName, String lastName,
                @JsonFormat(pattern = "yyyy-MM-dd") LocalDate birthdate,
                String email) {
