+++ src/main/java/pl/wsb/fitnesstracker/user/api/UserMapper.java
@@ -1,13 +1,11 @@
-package pl.wsb.fitnesstracker.user.internal;
+package pl.wsb.fitnesstracker.user.api;
 
 import org.springframework.stereotype.Component;
-import pl.wsb.fitnesstracker.user.api.User;
-import pl.wsb.fitnesstracker.user.api.UserEmailDto;
 
 @Component
-class UserMapper {
+public class UserMapper {
 
-    UserDto toDto(User user) {
+    public UserDto toDto(User user) {
         return new UserDto(user.getId(),
                 user.getFirstName(),
                 user.getLastName(),
@@ -15,7 +13,7 @@ class UserMapper {
                 user.getEmail());
     }
 
-    User toEntity(UserDto userDto) {
+    public User toEntity(UserDto userDto) {
         return new User(
                 userDto.firstName(),
                 userDto.lastName(),
@@ -23,8 +21,7 @@ class UserMapper {
                 userDto.email());
     }
 
-    UserEmailDto toEmailDto(User user) {
+    public UserEmailDto toEmailDto(User user) {
         return new UserEmailDto(user.getId(), user.getEmail());
     }
-
 }
